requirements/requirements.txt
Django==1.11.1
django-filter==1.0.3
djangorestframework==3.6.3
Markdown==2.6.8
